server/requirements.txt
zappa
flask
flask-graphql
graphene
graphene_django
geohash2
gpxpy
pykml
bs4
dj_database_url
psycopg2-binary
shortuuid
jsonfield
sendwithus
oauth2client
-e git://github.com/joram/stravalib.git#egg=stravalib
-e git://github.com/wckd/python-yr.git#egg=python-yr
